requirements.txt
kafka-python==2.0.2
Flask==2.2.2
dacite==1.7.0
redis==3.5.3
flasgger==0.9.5